requirements.txt
pandas
numpy 
matplotlib
Pillow
torch
flopth